Enabled optimization and measured execution time for Attitude controller
Attitude controller takes approximately 0.9 ms to execute

+++ KugleFirmware/Libraries/Modules/Estimators/COMEKF/COMEKF.h
@@ -40,7 +40,7 @@ class COMEKF
 	private:
 		Parameters& _params;
 		Timer * _microsTimer;
-		uint16_t _prevTimerValue;
+		uint32_t _prevTimerValue;
 
 		int32_t _prevVelocity[2];
