+++ KugleFirmware/Libraries/Modules/Estimators/VelocityEKF/VelocityEKF.h
@@ -41,7 +41,7 @@ class VelocityEKF
 	private:
 		Parameters& _params;
 		Timer * _microsTimer;
-		uint16_t _prevTimerValue;
+		uint32_t _prevTimerValue;
 
 		int32_t _prevEncoderTicks[3];
